+++ Client/Models/Sklad_rashod_tov.cs
@@ -12,7 +12,7 @@ namespace Client.Models
     {
         [Key]
         public decimal kod_zap { get; set; }
-        public Nullable<decimal> kod_rashoda { get; set; }
+        public Nullable<int> kod_rashoda { get; set; }
         public Nullable<decimal> tov { get; set; }
         public Nullable<decimal> count { get; set; }
         public Nullable<decimal> count_pachek { get; set; }
